Create licence
<b>Attribution-NonCommercial-NoDerivatives 4.0 International (CC BY-NC-ND 4.0)</b>
<br><b>You are free to:</b>
<br>Share — copy and redistribute the material in any medium or format. Under the following terms:
Attribution — You must give appropriate credit, provide a link to the license, and indicate if changes were made. You may do so in any reasonable manner, but not in any way that suggests the licensor endorses you or your use.

NonCommercial — You may not use the material for commercial purposes.

NoDerivatives — If you remix, transform, or build upon the material, you may not distribute the modified material.